Fix bot getting stuck in a loop
