+++ main.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/diamondburned/arikawa/v3/api"
 	"github.com/diamondburned/arikawa/v3/discord"
 	"github.com/diamondburned/arikawa/v3/gateway"
 	"github.com/diamondburned/arikawa/v3/session"
@@ -28,6 +27,13 @@ func main() {
 	queue := make(chan message)
 
 	s.AddHandler(func(c *gateway.MessageCreateEvent) {
+		me, err := s.Me()
+		if err != nil {
+			log.Printf("error finding myself: %v", err)
+		}
+		if me.ID == c.Author.ID {
+			return
+		}
 		if replacers := findReplacers(c.Message); len(replacers) > 0 {
 			queue <- message{
 				content:   c.Message,
@@ -70,20 +76,6 @@ func startHealthCheck() {
 
 func getHealth(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
-}
-
-func hideEmbeds(s *session.Session, m message) {
-	oldFlags := m.content.Flags
-	newFlags := oldFlags | discord.SuppressEmbeds
-
-	editMsgData := api.EditMessageData{
-		Flags: &newFlags,
-	}
-
-	_, err := s.EditMessageComplex(m.content.ChannelID, m.content.ID, editMsgData)
-	if err != nil {
-		log.Println("error editing message:", err)
-	}
 }
 
 type message struct {
